+++ modules/repository-azure/src/main/java/org/elasticsearch/repositories/azure/AzureStorageSettings.java
@@ -108,6 +108,10 @@ final class AzureStorageSettings {
     );
 
     private final String account;
+
+    @Nullable
+    private final String sasToken;
+
     private final String connectString;
     private final String endpointSuffix;
     private final TimeValue timeout;
@@ -126,6 +130,7 @@ final class AzureStorageSettings {
         Integer proxyPort
     ) {
         this.account = account;
+        this.sasToken = sasToken;
         this.connectString = buildConnectString(account, key, sasToken, endpointSuffix);
         this.endpointSuffix = endpointSuffix;
         this.timeout = timeout;
@@ -148,6 +153,10 @@ final class AzureStorageSettings {
                 throw new SettingsException("Azure proxy host is unknown.", e);
             }
         }
+    }
+
+    public String getSasToken() {
+        return sasToken;
     }
 
     public String getEndpointSuffix() {
