+++ modules/repository-azure/src/test/java/org/elasticsearch/repositories/azure/AzureBlobContainerRetriesTests.java
@@ -9,49 +9,27 @@ package org.elasticsearch.repositories.azure;
 
 import fixture.azure.AzureHttpHandler;
 
-import com.azure.storage.common.policy.RequestRetryOptions;
-import com.azure.storage.common.policy.RetryPolicyType;
-import com.sun.net.httpserver.HttpExchange;
 import com.sun.net.httpserver.HttpHandler;
-import com.sun.net.httpserver.HttpServer;
 
-import org.elasticsearch.cluster.metadata.RepositoryMetadata;
 import org.elasticsearch.common.Strings;
 import org.elasticsearch.common.UUIDs;
 import org.elasticsearch.common.blobstore.BlobContainer;
-import org.elasticsearch.common.blobstore.BlobPath;
 import org.elasticsearch.common.bytes.BytesReference;
 import org.elasticsearch.common.io.Streams;
 import org.elasticsearch.common.lucene.store.ByteArrayIndexInput;
 import org.elasticsearch.common.lucene.store.InputStreamIndexInput;
-import org.elasticsearch.common.network.InetAddresses;
-import org.elasticsearch.common.settings.MockSecureSettings;
-import org.elasticsearch.common.settings.Settings;
 import org.elasticsearch.common.unit.ByteSizeUnit;
-import org.elasticsearch.common.unit.ByteSizeValue;
-import org.elasticsearch.common.util.BigArrays;
 import org.elasticsearch.common.util.concurrent.CountDown;
 import org.elasticsearch.core.SuppressForbidden;
-import org.elasticsearch.core.TimeValue;
-import org.elasticsearch.core.Tuple;
-import org.elasticsearch.mocksocket.MockHttpServer;
 import org.elasticsearch.rest.RestStatus;
 import org.elasticsearch.rest.RestUtils;
-import org.elasticsearch.test.ESTestCase;
-import org.elasticsearch.threadpool.TestThreadPool;
-import org.elasticsearch.threadpool.ThreadPool;
-import org.junit.After;
-import org.junit.Before;
 
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.InputStreamReader;
-import java.net.InetAddress;
-import java.net.InetSocketAddress;
 import java.nio.file.NoSuchFileException;
 import java.util.Arrays;
-import java.util.Base64;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Locale;
@@ -59,22 +37,11 @@ import java.util.Map;
 import java.util.Objects;
 import java.util.Optional;
 import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.concurrent.atomic.AtomicLong;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 
 import static java.nio.charset.StandardCharsets.UTF_8;
-import static org.elasticsearch.repositories.azure.AzureRepository.Repository.CONTAINER_SETTING;
-import static org.elasticsearch.repositories.azure.AzureRepository.Repository.LOCATION_MODE_SETTING;
-import static org.elasticsearch.repositories.azure.AzureRepository.Repository.MAX_SINGLE_PART_UPLOAD_SIZE_SETTING;
-import static org.elasticsearch.repositories.azure.AzureStorageSettings.ACCOUNT_SETTING;
-import static org.elasticsearch.repositories.azure.AzureStorageSettings.ENDPOINT_SUFFIX_SETTING;
-import static org.elasticsearch.repositories.azure.AzureStorageSettings.KEY_SETTING;
-import static org.elasticsearch.repositories.azure.AzureStorageSettings.MAX_RETRIES_SETTING;
-import static org.elasticsearch.repositories.azure.AzureStorageSettings.TIMEOUT_SETTING;
 import static org.elasticsearch.repositories.blobstore.ESBlobStoreRepositoryIntegTestCase.randomBytes;
 import static org.hamcrest.Matchers.containsString;
 import static org.hamcrest.Matchers.equalTo;
@@ -87,102 +54,7 @@ import static org.hamcrest.Matchers.lessThanOrEqualTo;
  * This class tests how a {@link AzureBlobContainer} and its underlying SDK client are retrying requests when reading or writing blobs.
  */
 @SuppressForbidden(reason = "use a http server")
-public class AzureBlobContainerRetriesTests extends ESTestCase {
-
-    private static final long MAX_RANGE_VAL = Long.MAX_VALUE - 1L;
-
-    private HttpServer httpServer;
-    private HttpServer secondaryHttpServer;
-    private ThreadPool threadPool;
-    private AzureClientProvider clientProvider;
-
-    @Before
-    public void setUp() throws Exception {
-        threadPool = new TestThreadPool(
-            getTestClass().getName(),
-            AzureRepositoryPlugin.executorBuilder(),
-            AzureRepositoryPlugin.nettyEventLoopExecutorBuilder(Settings.EMPTY)
-        );
-        httpServer = MockHttpServer.createHttp(new InetSocketAddress(InetAddress.getLoopbackAddress(), 0), 0);
-        httpServer.start();
-        secondaryHttpServer = MockHttpServer.createHttp(new InetSocketAddress(InetAddress.getLoopbackAddress(), 0), 0);
-        secondaryHttpServer.start();
-        clientProvider = AzureClientProvider.create(threadPool, Settings.EMPTY);
-        clientProvider.start();
-        super.setUp();
-    }
-
-    @After
-    public void tearDown() throws Exception {
-        clientProvider.close();
-        httpServer.stop(0);
-        secondaryHttpServer.stop(0);
-        super.tearDown();
-        ThreadPool.terminate(threadPool, 10L, TimeUnit.SECONDS);
-    }
-
-    private BlobContainer createBlobContainer(final int maxRetries) {
-        return createBlobContainer(maxRetries, null, LocationMode.PRIMARY_ONLY);
-    }
-
-    private BlobContainer createBlobContainer(final int maxRetries, String secondaryHost, final LocationMode locationMode) {
-        final Settings.Builder clientSettings = Settings.builder();
-        final String clientName = randomAlphaOfLength(5).toLowerCase(Locale.ROOT);
-
-        String endpoint = "ignored;DefaultEndpointsProtocol=http;BlobEndpoint=" + getEndpointForServer(httpServer, "account");
-        if (secondaryHost != null) {
-            endpoint += ";BlobSecondaryEndpoint=" + getEndpointForServer(secondaryHttpServer, "account");
-        }
-        clientSettings.put(ENDPOINT_SUFFIX_SETTING.getConcreteSettingForNamespace(clientName).getKey(), endpoint);
-        clientSettings.put(MAX_RETRIES_SETTING.getConcreteSettingForNamespace(clientName).getKey(), maxRetries);
-        clientSettings.put(TIMEOUT_SETTING.getConcreteSettingForNamespace(clientName).getKey(), TimeValue.timeValueMillis(500));
-
-        final MockSecureSettings secureSettings = new MockSecureSettings();
-        secureSettings.setString(ACCOUNT_SETTING.getConcreteSettingForNamespace(clientName).getKey(), "account");
-        final String key = Base64.getEncoder().encodeToString(randomAlphaOfLength(14).getBytes(UTF_8));
-        secureSettings.setString(KEY_SETTING.getConcreteSettingForNamespace(clientName).getKey(), key);
-        clientSettings.setSecureSettings(secureSettings);
-
-        final AzureStorageService service = new AzureStorageService(clientSettings.build(), clientProvider) {
-            @Override
-            RequestRetryOptions getRetryOptions(LocationMode locationMode, AzureStorageSettings azureStorageSettings) {
-                return new RequestRetryOptions(
-                    RetryPolicyType.EXPONENTIAL,
-                    maxRetries + 1,
-                    60,
-                    50L,
-                    100L,
-                    // The SDK doesn't work well with ip endponts. Secondary host endpoints that contain
-                    // a path causes the sdk to rewrite the endpoint with an invalid path, that's the reason why we provide just the host +
-                    // port.
-                    secondaryHost != null ? secondaryHost.replaceFirst("/account", "") : null
-                );
-            }
-
-            @Override
-            long getUploadBlockSize() {
-                return ByteSizeUnit.MB.toBytes(1);
-            }
-
-            @Override
-            int getMaxReadRetries(String clientName) {
-                return maxRetries;
-            }
-        };
-
-        final RepositoryMetadata repositoryMetadata = new RepositoryMetadata(
-            "repository",
-            AzureRepository.TYPE,
-            Settings.builder()
-                .put(CONTAINER_SETTING.getKey(), "container")
-                .put(ACCOUNT_SETTING.getKey(), clientName)
-                .put(LOCATION_MODE_SETTING.getKey(), locationMode)
-                .put(MAX_SINGLE_PART_UPLOAD_SIZE_SETTING.getKey(), new ByteSizeValue(1, ByteSizeUnit.MB))
-                .build()
-        );
-
-        return new AzureBlobContainer(BlobPath.EMPTY, new AzureBlobStore(repositoryMetadata, service, BigArrays.NON_RECYCLING_INSTANCE));
-    }
+public class AzureBlobContainerRetriesTests extends AbstractAzureServerTestCase {
 
     public void testReadNonexistentBlobThrowsNoSuchFileException() {
         final BlobContainer blobContainer = createBlobContainer(between(1, 5));
@@ -595,55 +467,6 @@ public class AzureBlobContainerRetriesTests extends ESTestCase {
             // It does round robin, first tries on the primary, then on the secondary
             assertThat(failedHeadCalls.get(), equalTo(1));
             assertThat(failedGetCalls.get(), equalTo(1));
-        }
-    }
-
-    private static byte[] randomBlobContent() {
-        return randomByteArrayOfLength(randomIntBetween(1, 1 << 20)); // rarely up to 1mb
-    }
-
-    private static final Pattern RANGE_PATTERN = Pattern.compile("^bytes=([0-9]+)-([0-9]+)$");
-
-    private static Tuple<Long, Long> getRanges(HttpExchange exchange) {
-        final String rangeHeader = exchange.getRequestHeaders().getFirst("X-ms-range");
-        if (rangeHeader == null) {
-            return Tuple.tuple(0L, MAX_RANGE_VAL);
-        }
-
-        final Matcher matcher = RANGE_PATTERN.matcher(rangeHeader);
-        assertTrue(rangeHeader + " matches expected pattern", matcher.matches());
-        final long rangeStart = Long.parseLong(matcher.group(1));
-        final long rangeEnd = Long.parseLong(matcher.group(2));
-        assertThat(rangeStart, lessThanOrEqualTo(rangeEnd));
-        return Tuple.tuple(rangeStart, rangeEnd);
-    }
-
-    private static int getRangeStart(HttpExchange exchange) {
-        return Math.toIntExact(getRanges(exchange).v1());
-    }
-
-    private static Optional<Integer> getRangeEnd(HttpExchange exchange) {
-        final long rangeEnd = getRanges(exchange).v2();
-        if (rangeEnd == MAX_RANGE_VAL) {
-            return Optional.empty();
-        }
-        return Optional.of(Math.toIntExact(rangeEnd));
-    }
-
-    private String getEndpointForServer(HttpServer server, String accountName) {
-        InetSocketAddress address = server.getAddress();
-        return "http://" + InetAddresses.toUriString(address.getAddress()) + ":" + address.getPort() + "/" + accountName;
-    }
-
-    private void readFromInputStream(InputStream inputStream, long bytesToRead) {
-        try {
-            long totalBytesRead = 0;
-            while (inputStream.read() != -1 && totalBytesRead < bytesToRead) {
-                totalBytesRead += 1;
-            }
-            assertThat(totalBytesRead, equalTo(bytesToRead));
-        } catch (IOException e) {
-            throw new RuntimeException(e);
         }
     }
 }
